# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the workdadventure back office written at the Hochschule Bochum",
        "title": "Workadventure Admin Back Office API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "GitHub Issues",
            "url": "https://github.com/bo-mathadventure/admin"
        },
        "license": {
            "name": "AGPL 3.0",
            "url": "https://github.com/teamdigitale/licenses/blob/master/AGPL-3.0-or-later"
        },
        "version": "1.0"
    },
    "host": "localhost:4664",
    "basePath": "/",
    "paths": {
...
        "/system/admin/ban": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all bans. Requires permission admin.ban.view or admin.ban.create or admin.ban.delete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List bans",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/admin.AdminBanResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a ban for an identifier or ip address. Requires permission admin.ban.create",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create ban",
                "parameters": [
                    {
                        "description": "-",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.CreateBan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.AdminBanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/system/admin/ban/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get ban by ID. Requires permission admin.ban.view or admin.ban.create or admin.ban.delete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get specific ban",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ban ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.AdminBanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete ban by ID. Sets validUntil to now() to keep history. Requires permission admin.ban.delete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete ban",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ban ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.AdminBanResponse": {
            "type": "object",
            "properties": {
                "check": {
                    "type": "string",
                    "example": "60948703-fca9-4491-b3bc-588188d93eb3"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05Z07:00"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "banned for verbal abuse"
                },
                "validUntil": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05Z07:00"
                }
            }
        },
...
        "admin.CreateBan": {
            "type": "object",
            "properties": {
                "check": {
                    "type": "string",
                    "example": "60948703-fca9-4491-b3bc-588188d93eb3"
                },
                "message": {
                    "type": "string",
                    "example": "banned for verbal abuse"
                },
                "validUntil": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05Z07:00"
                }
            }
        },
...
        "handler.APIResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "extra": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "ERR_INVALID_PERMISSIONS"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT user token from login",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}